Add operators

+++ common/string_arg.h
@@ -1,16 +1,17 @@
 #ifndef STRING_ARG_H
 #define STRING_ARG_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 typedef struct StringArg{
     char* value;
-    uint8_t len;
+    size_t len;
 } string_arg;
 
 typedef struct StringArgs{
     string_arg* args;
-    uint8_t len;
+    size_t len;
 } string_args;
 
 #endif
